requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
streamlit==1.28.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.1.4
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
aiofiles==23.2.1
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
aiosqlite==0.19.0
plotly==5.17.0
